amountChooser

+++ appsDiscovery/src/main/java/org/kinecosystem/appsdiscovery/sender/service/SendKinServiceBase.java
@@ -6,14 +6,12 @@ import android.os.Binder;
 import android.os.IBinder;
 import android.support.annotation.NonNull;
 
-import java.math.BigDecimal;
-
 public abstract class SendKinServiceBase extends Service {
 
     //Developer needs to implements
     //these is called from background thread so can be called async
     public abstract @NonNull
-    KinTransferComplete transferKin(@NonNull final String toAddress, final BigDecimal amount, @NonNull final String memo) throws KinTransferException;
+    KinTransferComplete transferKin(@NonNull final String toAddress, final int amount, @NonNull final String memo) throws KinTransferException;
 
     public abstract int getCurrentBalance() throws BalanceException;
